Test uncompress code. Moved more stuff to Node.cs from Networker.cs. Started integrating things.

+++ GUI_FrontEnd/MainForm.cs
@@ -77,7 +77,7 @@ namespace GUI_FrontEnd
 
         private void maxBackupSupport_ValueChanged(object sender, EventArgs e)
         {
-            Settings.Default.maxBackupCapacity = numUpDownMaxBackupCapacity.Value;
+            Settings.Default.maxBackupCapacity = (long)numUpDownMaxBackupCapacity.Value;
         }    
 
         private void tabControl_SelectedIndexChanged(object sender, EventArgs e)
